requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-dht==4.0.5
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
mysql-connector-python==9.1.0
pigpio==1.78
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
smbus==1.1.post2
smbus2==0.5.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
